Make sure we clean daily data living in different path
We used to have hourly and daily at the same path (called sourcePath)
before. Now they are separate and therefore the cleaner cannot find
daily files. As a precaution - it does not delete hourly files if there
the daily files are not found. This patch adds destSource which allows
us to have daily files in a different path.

Also added some logging to make it more clear of what is happening.